Refactor for Projects Controller and Service : Add Create and Update Task DTO

+++ src/main/java/com/taski/projects/service/ProjectService.java
@@ -7,9 +7,11 @@ import com.taski.projects.dto.DeleteProjectDTO;
 import com.taski.projects.dto.UpdateProjectDTO;
 import com.taski.projects.model.Project;
 import com.taski.projects.repository.ProjectsRepository;
+import com.taski.utils.ValidationUtils;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -23,17 +25,24 @@ public class ProjectService {
      }
 
     public void createProject(CreateProjectDTO projectDTO){
+         projectDTO.setUserId(ValidationUtils.getUserID());
          Project project = new Project(projectDTO.getUserId(), projectDTO.getName(), projectDTO.getDescription());
          projectsRepository.createProject(project);
     }
 
     public boolean updateProject(UpdateProjectDTO projectDTO){
+        if(!userHasProject(projectDTO.getId())){
+            throw new IllegalStateException("Current user has no access to this project.");
+        }
         Project project = new Project(projectDTO.getName(),projectDTO.getId(), projectDTO.getDescription());
         int rowsAffected = projectsRepository.updateProject(projectDTO);
         return rowsAffected > 0;
     }
 
     public void deleteProject(Long id){
+        if(!userHasProject(id)){
+            throw new IllegalStateException("Current user has no access to this project.");
+        }
          projectsRepository.deleteProject(id);
     }
 
@@ -46,6 +55,13 @@ public class ProjectService {
      User user = userOPT.get();
 
      return projectsRepository.getProjectsByUserId(user.getId());
-
     }
+
+    private boolean userHasProject(Long projectId) {
+        Long userId = ValidationUtils.getUserID();
+        return getProjectsByUserId(userId)
+                .stream()
+                .anyMatch(project -> Objects.equals(project.getId(), projectId));
+    }
+
 }
